requirements.txt
pydantic>=2
python-dotenv>=1.0.0
